added code to minimax
currently, it is just code that I made blind, and it is in need of helper functions to clean up the code a bit if reasonable. Other notes for future me: minimax currently will be returning a value move pair, and that will need to be dealt with when it is being called. value is changed when minimax is called inside itself to propagate values down.